requirements.txt
pandas
numpy
pyodbc
psycopg2-binary
sqlalchemy
openpyxl
fastapi
uvicorn
python-dotenv
